requirements.txt
imgaug==0.4.0
matplotlib==3.3.4
numpy==1.21.1
opencv-python-headless
Pillow==9.0.1
pytorch_lightning==1.5.10
streamlit==1.5.0
-f https://download.pytorch.org/whl/torch_stable.html                                                                                                                                                             
torch==1.9.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.0+cpu
